Take-skip rope
The task is to read and skip through a string, extracting a hidden message. First, we separate digits from symbols, then we make 2 lists of even_indexed and odd_indexed digits. Then we iterate over the string list and extract the number of symbols, according to the skip-and-take list of digits.